# Priority based action script loading