--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to this project will be documented in this file. (Notable changes means some internal changes may not be here!)
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog (1.1)](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning (2.0)](https://semver.org/spec/v2.0.0.html).
+
+## [unreleased]
+
+### Added
+- Higher priority for main action scripts loading ([#112](https://github.com/text-forge/text-forge/pull/112))
 
 ## [v0.2-beta1] - 2025-9-25