… core/main.tscn