<!--Provide a succinct and descriptive title for the pull request, e.g., "Improve caching mechanism for API calls"-->

## Type of Change
<!--Remove unrelated items-->
- UI/UX improvement

## Description
<!--Provide a detailed explanation of the changes you have made. Include the reasons behind these changes and any relevant context. Link any related issues.-->
Action Scripts are divided into two main categories (file menu) and secondary (the rest). The user interface will be fully usable after loading the main items, so there is no need to wait until the end of the loading process. This roughly halves the actual startup time (-60.4%), although the overall startup is slightly slower.

## Testing
<!--Detail the testing you have performed to ensure that these changes function as intended. Include information about any added tests.-->
Works perfectly.

## Impact
<!--Discuss the impact of your changes on the project. This might include effects on performance, new dependencies, or changes in behaviour.-->
### Environment
|Hardware|Information                        |Software  |information                       |
|--------|-----------------------------------|----------|----------------------------------|
|CPU     |Intel Core i5-10210U CPU @ 1.60GHz |OS        |Windows 11 24H2                   |
|RAM     |DDR4 32GB 2667 MT/s                |Power Mode|Best Performance                  |
|Disk    |SSD (NVMe)                         |Godot     |v4.5.stable.official [876b29033]  |
|GPU     |NVIDIA GeForce MX350 2GB           |Build     |editor                            |

### Before

|Param                                   |Unit|Min  |Max  |Avg   |Repeats|
|----------------------------------------|:--:|:---:|:---:|:----:|:-----:|
|Startup Time                            |msec|253  |276  |260.4 |5      |
|Panels Loading Time (background)        |msec|278  |295  |284.4 |5      |
|Action Scripts Loading Time (background)|msec|1205 |1513 |1390.0|5      |
|Total Startup Time (with background)    |msec|1458 |1771 |1650.0|5      |

### After

|Param                                        |Unit| Min | Max |  Avg  |Repeats|
|---------------------------------------------|:--:|:---:|:---:|:-----:|:-----:|
|Startup Time                                 |msec|239  |308  |275.6  |5      |
|Panels Loading Time (background)             |msec|269  |348  |300.6  |5      |
|Main Action Scripts Loading Time (background)|msec|349  |434  |376.8  |5      |
|Total Startup Time (with background)         |msec|603  |742  |652.6  |5      |
|All Action Scripts Loading Time (background) |msec|1163 |1567 |1360.4 |5      |
|Total Setup Time (with background)           |msec|1402 |1821 |1636.0 |5      |

### Overall Comparison
|Param                                        |     Change     |Percent|
|---------------------------------------------|:--------------:|:-----:|
|Startup Time                                 |260.4 -> 275.6  |+5.8%  |
|Panels Loading Time (background)             |284.4 -> 300.6  |+5.6%  |
|Main Action Scripts Loading Time (background)|376.8           |-------|
|Total Startup Time (with background)         |1650.0 -> 652.6 |-60.4% |
|All Action Scripts Loading Time (background) |1390.0 -> 1360.4|-2.0%  |
|Total Setup Time (with background)           |1636.0          |-------|

## Additional Information
<!--Any additional information that reviewers should be aware of.-->
This update includes test improvements, improvements in the behavior of automatically loading the last file, a separate loader for main Action Scripts, setting the default values of the View menu as the editor default, and removing the search_delay parameter from the performance test.

## Checklist
- [x] My code adheres to the coding and style guidelines of the project.
- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand areas.
- [x] I have made corresponding changes to the documentation.
- [x] My changes generate no new warnings
- [x] Changes were added to CHANGELOG


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- New Features
  - Code editor enhancements: minimap, highlight current line and occurrences, line length guides (80/100), and gutters for bookmarks, line numbers, and folding.
  - Faster, smoother startup via staged loading; options become available sooner.
  - Editor now auto-focuses after the last file auto-loads.

- Documentation
  - Changelog updated with specific spec references and a new Unreleased section.

- Tests
  - Performance tests expanded to separately measure startup and file-open paths with refined timing outputs.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->